Add files via upload
This script breaks down lyrics of all songs on Taylor Swift's new album and writes the count of each word to a CSV file for further analysis. You can plug in any txt file with words/strings and the script will still perform as intended as long as you change the input file name/location at the top of the script. Lyrics were copy and pasted from the internet into a text file